+++ src/world_loader.rs
@@ -17,27 +17,27 @@ pub struct JsonWorldData {
 enum JsonObject {
     Square {
         model: String,
-        origin: [f32; 3],
-        rotation_axis: [f32; 3],
-        rotation_angle: f32,
-        scale: f32
+        origin: [f64; 3],
+        rotation_axis: [f64; 3],
+        rotation_angle: f64,
+        scale: f64
     },
     Rectangle {
         model: String,
-        origin: [f32; 3],
-        rotation_axis: [f32; 3],
-        rotation_angle: f32,
-        width: f32,
-        height: f32,
+        origin: [f64; 3],
+        rotation_axis: [f64; 3],
+        rotation_angle: f64,
+        width: f64,
+        height: f64,
         color: [u8; 3]
     },
     SpinningCube {
         model: String,
-        origin: [f32; 3],
-        rotation_axis: [f32; 3],
-        rotation_angle: f32,
-        angular_velocity: f32,
-        scale: f32
+        origin: [f64; 3],
+        rotation_axis: [f64; 3],
+        rotation_angle: f64,
+        angular_velocity: f64,
+        scale: f64
     }
 }
 
@@ -46,7 +46,7 @@ enum JsonObject {
 #[allow(dead_code)]
 enum JsonLight {
     Point {
-        origin: [[f32; 3]; 3]
+        origin: [[f64; 3]; 3]
     }
 }
 
@@ -69,11 +69,11 @@ pub fn load_world<'a>(world_path: &'a str, model_loader: &'a ModelLoader) -> Vec
                 } => {
                     world_objects::Entity::Square(world_objects::Square {
                         model: &model_loader.get_model(model),
-                        origin: Point3::<f32>::new(origin[0], origin[1], origin[2]),
-                        rotation: Rotation3::<f32>::from_axis_angle(
+                        origin: Point3::<f64>::new(origin[0], origin[1], origin[2]),
+                        rotation: Rotation3::<f64>::from_axis_angle(
                             &Unit::new_normalize(Vector3::new(rotation_axis[0], rotation_axis[1], rotation_axis[2])), 
                             *rotation_angle),
-                        scale: Matrix4::<f32>::new_scaling(*scale)
+                        scale: Matrix4::<f64>::new_scaling(*scale)
                     })
                 },                
                 JsonObject::Rectangle {
@@ -87,11 +87,11 @@ pub fn load_world<'a>(world_path: &'a str, model_loader: &'a ModelLoader) -> Vec
                 } => {
                     world_objects::Entity::Rectangle(world_objects::Rectangle {
                         model: &model_loader.get_model(model),
-                        origin: Point3::<f32>::new(origin[0], origin[1], origin[2]),
-                        rotation: Rotation3::<f32>::from_axis_angle(
+                        origin: Point3::<f64>::new(origin[0], origin[1], origin[2]),
+                        rotation: Rotation3::<f64>::from_axis_angle(
                             &Unit::new_normalize(Vector3::new(rotation_axis[0], rotation_axis[1], rotation_axis[2])), 
                             *rotation_angle),
-                        scale: Matrix4::<f32>::new_nonuniform_scaling(&Vector3::new(*width, *height, 1.0)),
+                        scale: Matrix4::<f64>::new_nonuniform_scaling(&Vector3::new(*width, *height, 1.0)),
                         color: Color::new(color[0], color[1], color[2])
                     })
                 },
@@ -105,11 +105,11 @@ pub fn load_world<'a>(world_path: &'a str, model_loader: &'a ModelLoader) -> Vec
                 } => {
                     world_objects::Entity::SpinningCube(world_objects::SpinningCube {
                         model: &model_loader.get_model(model),
-                        origin: Point3::<f32>::new(origin[0], origin[1], origin[2]),
-                        rotation_axis: Vector3::<f32>::new(rotation_axis[0], rotation_axis[1], rotation_axis[2]),
+                        origin: Point3::<f64>::new(origin[0], origin[1], origin[2]),
+                        rotation_axis: Vector3::<f64>::new(rotation_axis[0], rotation_axis[1], rotation_axis[2]),
                         rotation_angle: *rotation_angle,
                         rotation_velocity: *angular_velocity,
-                        scale: Matrix4::<f32>::new_scaling(*scale)
+                        scale: Matrix4::<f64>::new_scaling(*scale)
                     })
                 },
             }
